modified autoFollow.pl

+++ autoFollow.pl
@@ -1,17 +1,51 @@
 #!/usr/bin/perl
+
+=head1 NAME
+
+auto follow back / remove
+
+=head1 SYNOPSIS
+
+randomPost [options] 
+
+ Options:
+   --dry-run         dryrun
+   --users           number of users per 1time execution
+=cut
+
 use strict;
+use warnings;
+
 use Net::Twitter;
 use Data::Dumper;
 use utf8;
 use Encode;
+use Pod::Usage;
+use Getopt::Long qw/:config posix_default no_ignore_case bundling auto_help/;
+
+
+GetOptions(
+	\my %opts, qw/
+	dry-run
+	users=i
+	/
+	) or pod2usage(verbose => 0);
 
 use update;
+
+my $users = 10; #1回につき10人まで
+
+if($opts{'users'}){
+	$users = $opts{'users'};
+}
+
 
 my $r_following  = myUpdateStatus::followingIds();
 my $r_followers =  myUpdateStatus::followersIds();
 
 #my @following = sort {$a->{'id'} <=> $b->{'id'}} @$r_following;
 #my @followers = sort {$a->{'id'} <=> $b->{'id'}} @$r_followers;
+print Dumper $r_following;
 
 my @following = sort{$a <=> $b} @$r_following;
 my @followers = sort{$a <=> $b} @$r_followers;
@@ -65,14 +99,13 @@ while(1){
 if(@remove){
 
 	foreach(@remove){
-		my $removed = myUpdateStatus::destroyFriend($_);
-###			if(defined($removed)){
-###				print "$_をリムーブしました\n";
-###			}
-###			else{
-###				print "$_のリムーブに失敗しました。\n"
-###			}
-			sleep(5);
+		if(!$opts{'dry-run'}){
+			my $removed = myUpdateStatus::destroyFriend($_);
+		}
+		else{
+			print "$_ is removed\n";
+		}
+		sleep(5);
 	}
 }
 
@@ -81,16 +114,15 @@ if(@follow){
 		#スパムかどうかチェックする
 		#		my $user = isSpam($_);
 		###		if(!defined($user)){
-		my $followed = myUpdateStatus::createFriend($_);
-		###			if(defined($followed)){
-		###				print "$_をフォローしました\n";
-		###			}
-		###			else{
-		###				print "$_のフォローに失敗しました。\n"
-		###			}
+		if(!$opts{'dry-run'}){
+			my $followed = myUpdateStatus::createFriend($_);
+		}
+		else{
+			print "$_ is followed\n";
+		}
 		sleep(5);
 	}
 }
 
-
+exit;
 #print "終了\n";
